feat: add published multiplatform images
