+++ .github/workflows/commit.yml
@@ -9,8 +9,12 @@ on:
     branches:
       - main
 
+defaults:
+  run:
+    shell: 'bash --noprofile --norc -Eeuo pipefail {0}'
+
 env:
-  python_version: "3.10"
+  python_version: "3.11"
 
 jobs:
   lint:
@@ -40,9 +44,11 @@ jobs:
           path: ~/.local/share/virtualenvs
           key: ${{ runner.os }}-python-${{ env.python_version }}-pipenv-${{ hashFiles('Pipfile.lock') }}
       - name: Install the dependencies
-        run: |
-          python -m pip install --upgrade pipenv
-          pipenv install --python ${{ env.python_version }} --deploy --ignore-pipfile --dev
+        run: python -m pip install --upgrade pipenv
+      - name: Install Task
+        uses: arduino/setup-task@v1
+      - name: Initialize the repo
+        run: task -v init
       - name: Gather the image matrix
         id: set-image-outputs
         run: |
@@ -76,17 +82,26 @@ jobs:
       - name: Install the dependencies
         run: |
           python -m pip install --upgrade pipenv
-          pipenv install --python ${{ env.python_version }} --deploy --ignore-pipfile --dev
           mkdir "${RUNNER_TEMP}/bin"
           curl -sSfL https://raw.githubusercontent.com/anchore/syft/main/install.sh | sh -s -- -b "${RUNNER_TEMP}/bin"
           chmod +x "${RUNNER_TEMP}/bin/syft"
           curl -sSfL https://raw.githubusercontent.com/anchore/grype/main/install.sh | sh -s -- -b "${RUNNER_TEMP}/bin"
           chmod +x "${RUNNER_TEMP}/bin/grype"
           echo "${RUNNER_TEMP}/bin" >> "${GITHUB_PATH}"
+      - name: Install Task
+        uses: arduino/setup-task@v1
+      - name: Initialize the repo
+        run: task -v init
       - name: Build the image
-        run: pipenv run invoke build --tool=${{ matrix.tool }} --environment=${{ matrix.environment }}
+        run: task -v build
+        env:
+          TOOL: ${{ matrix.tool }}
+          ENVIRONMENT: ${{ matrix.environment }}
       - name: Generate the SBOM
-        run: pipenv run invoke sbom --tool=${{ matrix.tool }} --environment=${{ matrix.environment }}
+        run: task -v sbom
+        env:
+          TOOL: ${{ matrix.tool }}
+          ENVIRONMENT: ${{ matrix.environment }}
       - name: Upload the SBOM
         uses: actions/upload-artifact@v3
         with:
@@ -94,7 +109,10 @@ jobs:
           path: sbom.*.json
           if-no-files-found: error
       - name: Generate Vuln scan results
-        run: pipenv run invoke vulnscan --tool=${{ matrix.tool }} --environment=${{ matrix.environment }}
+        run: task -v vulnscan
+        env:
+          TOOL: ${{ matrix.tool }}
+          ENVIRONMENT: ${{ matrix.environment }}
       - name: Upload Vuln scan result
         uses: actions/upload-artifact@v3
         with:
@@ -102,9 +120,12 @@ jobs:
           path: vulns.*.json
           if-no-files-found: error
       - name: Run tests
-        run: |
-          find tests -mindepth 1 -type d -exec chmod o+w {} \;
-          pipenv run invoke test --tool=${{ matrix.tool }} --environment=${{ matrix.environment }} --user=${{ matrix.user }} --debug
+        run: task test
+        env:
+          TOOL: ${{ matrix.tool }}
+          ENVIRONMENT: ${{ matrix.environment }}
+          USER: ${{ matrix.user }}
+          DEBUG: true
   bump-version:
     name: Bump version
     needs: [lint]
@@ -128,15 +149,15 @@ jobs:
           path: ~/.local/share/virtualenvs
           key: ${{ runner.os }}-python-${{ env.python_version }}-pipenv-${{ hashFiles('Pipfile.lock') }}
       - name: Install the dependencies
-        run: |
-          python -m pip install --upgrade pipenv
-          pipenv install --python ${{ env.python_version }} --deploy --ignore-pipfile --dev
+        run: python -m pip install --upgrade pipenv
+      - name: Install Task
+        uses: arduino/setup-task@v1
+      - name: Initialize the repo
+        run: task -v init
       - name: Bump the version
         id: bump-version
         run: |
-          git config --global user.name 'Seiso Automation'
-          git config --global user.email 'automation@seisollc.com'
-          pipenv run invoke release
+          task -v release
           GIT_TAG="$(git describe --tags)"
           BRANCH="$(git branch --show-current)"
           git push --atomic origin "${BRANCH}" "${GIT_TAG}"
@@ -166,17 +187,26 @@ jobs:
       - name: Install the dependencies
         run: |
           python -m pip install --upgrade pipenv
-          pipenv install --python ${{ env.python_version }} --deploy --ignore-pipfile --dev
           mkdir "${RUNNER_TEMP}/bin"
           curl -sSfL https://raw.githubusercontent.com/anchore/syft/main/install.sh | sh -s -- -b "${RUNNER_TEMP}/bin"
           chmod +x "${RUNNER_TEMP}/bin/syft"
           curl -sSfL https://raw.githubusercontent.com/anchore/grype/main/install.sh | sh -s -- -b "${RUNNER_TEMP}/bin"
           chmod +x "${RUNNER_TEMP}/bin/grype"
           echo "${RUNNER_TEMP}/bin" >> "${GITHUB_PATH}"
+      - name: Install Task
+        uses: arduino/setup-task@v1
+      - name: Initialize the repo
+        run: task -v init
       - name: Build the image
-        run: pipenv run invoke build --tool=${{ matrix.tool }} --environment=${{ matrix.environment }}
+        run: task -v build
+        env:
+          TOOL: ${{ matrix.tool }}
+          ENVIRONMENT: ${{ matrix.environment }}
       - name: Generate the SBOM
-        run: pipenv run invoke sbom --tool=${{ matrix.tool }} --environment=${{ matrix.environment }}
+        run: task -v sbom
+        env:
+          TOOL: ${{ matrix.tool }}
+          ENVIRONMENT: ${{ matrix.environment }}
       - name: Upload the SBOM
         uses: actions/upload-artifact@v3
         with:
@@ -184,7 +214,10 @@ jobs:
           path: sbom.*.json
           if-no-files-found: error
       - name: Generate Vuln scan results
-        run: pipenv run invoke vulnscan --tool=${{ matrix.tool }} --environment=${{ matrix.environment }}
+        run: task -v vulnscan
+        env:
+          TOOL: ${{ matrix.tool }}
+          ENVIRONMENT: ${{ matrix.environment }}
       - name: Upload Vuln scan result
         uses: actions/upload-artifact@v3
         with:
@@ -192,17 +225,22 @@ jobs:
           path: vulns.*.json
           if-no-files-found: error
       - name: Run tests
-        run: |
-          find tests -mindepth 1 -type d -exec chmod o+w {} \;
-          # We only test with the easy_infra user here; all supported users are tested via the PR workflow
-          pipenv run invoke test --tool=${{ matrix.tool }} --environment=${{ matrix.environment }} --user=easy_infra --debug
+        run: task test
+        env:
+          TOOL: ${{ matrix.tool }}
+          ENVIRONMENT: ${{ matrix.environment }}
+          USER: ${{ matrix.user }}
+          DEBUG: true
       - name: Log in to Docker Hub
         uses: docker/login-action@v2
         with:
           username: ${{ secrets.DOCKERHUB_USERNAME }}
           password: ${{ secrets.DOCKERHUB_TOKEN }}
       - name: Publish the image to Docker Hub
-        run: pipenv run invoke publish --tool=${{ matrix.tool }} --environment=${{ matrix.environment }}
+        run: task -v publish
+        env:
+          TOOL: ${{ matrix.tool }}
+          ENVIRONMENT: ${{ matrix.environment }}
       - name: Install cosign
         uses: sigstore/cosign-installer@main
       - name: Sign the image
