+++ .github/workflows/update.yml
@@ -6,13 +6,17 @@ on:
     - cron: '17 4 * * 1,4'
   workflow_dispatch:
 
+defaults:
+  run:
+    shell: 'bash --noprofile --norc -Eeuo pipefail {0}'
+
 env:
-  python_version: "3.10"
+  python_version: "3.11"
 
 jobs:
   update:
     name: Update
-    runs-on: ubuntu-20.04
+    runs-on: ubuntu-22.04
     steps:
       - name: Checkout the repository
         uses: actions/checkout@v3
@@ -20,11 +24,13 @@ jobs:
         with:
           python-version: ${{ env.python_version }}
       - name: Install the dependencies
-        run: |
-          python -m pip install --upgrade pipenv
-          pipenv install --deploy --ignore-pipfile --dev
+        run: python -m pip install --upgrade pipenv
+      - name: Install Task
+        uses: arduino/setup-task@v1
+      - name: Initialize the repo
+        run: task -v init
       - name: Update the repository
-        run: pipenv run invoke update
+        run: task -v update
       - name: Create or update a pull request
         uses: peter-evans/create-pull-request@v4
         with:
