+++ Pipfile
@@ -18,4 +18,4 @@ requests = "*"
 sphinx = "*"
 
 [requires]
-python_version = "3.10"
+python_version = "3.11"
