+++ easy_infra.yml
@@ -49,7 +49,7 @@ packages:
       environments:
       - none
       name: cloudformation
-    version: 2.13.7
+    version: 2.13.8
     version_argument: --version
   azure-cli:
     aliases:
@@ -57,12 +57,12 @@ packages:
     version: 2.51.0-1~focal
     version_argument: version
   checkov:
-    version: 2.3.358
+    version: 2.3.361
     version_argument: --version
   consul-template:
     helper:
     - all
-    version: v0.32.0
+    version: v0.33.0
     version_argument: --version
   envconsul:
     helper:
@@ -83,7 +83,7 @@ packages:
       env_vars:
       - TF_DATA_DIR
     security: *id003
-    version: 1.5.4
+    version: 1.5.5
     version_argument: version
   terratag:
     helper:
