+++ tests/terraform/hooks/secure_builtin_version/secure.tf
@@ -1,3 +1,3 @@
 terraform {
-  required_version = "1.5.4"
+  required_version = "1.5.5"
 }
